Create Roblox pass the bomb! script
This script will only teleport to the closest player once someone passes the bomb to you or when the bomb spawns on you. The reason why I chose the closest player instead of the farthest player is because the farthest the player is, the bomb won't be able to pass to them, it's like it is glitching out or there is an anti-teleport system that only detects if the player teleport within a few distance, that is when the anti teleport system kick in.